+++ app/Http/Controllers/Api/PackagesController.php
@@ -20,33 +20,60 @@ class PackagesController extends Controller
 
     public function store(Request $request)
     {
+        $validated = $request->validate(
+            [
+                'name' => 'required',
+                'price' => 'required',
+                'total_seats'=> 'required'
+
+            ]
+        );
+
         $data = Package::create($request->all());
         return response()->json([
             'status' => 'success',
             'message' => 'Data berhasil ditambahkan',
             'data' => $data
-        ]);
+        ], 200);
     }
 
     public function update(Request $request, $id)
     {
+
         $data = Package::find($id);
-        $data->update($request->all());
+
+        if($data)
+            {
+            $data->update($request->all());
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Data berhasil diupdate',
+                'data' => $data
+            ]);
+            }
         return response()->json([
-            'status' => 'success',
-            'message' => 'Data berhasil diupdate',
-            'data' => $data
-        ]);
+            'status'=> 'error',
+            'message'=> 'Package tidak ditemukan'
+        ],404);
+
     }
 
     public function destroy($id)
     {
         $data = Package::find($id);
-        $data->delete();
+
+        if ($data) {
+            $data->delete();
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Data berhasil dihapus',
+                'data' => $data
+            ]);
+        }
+
         return response()->json([
-            'status' => 'success',
-            'message' => 'Data berhasil dihapus',
-            'data' => $data
-        ]);
+            'status' => 'error',
+            'message' => 'Package tidak ditemukan'
+        ], 404);
     }
 }
